+++ activerecord/test/cases/adapters/postgresql/geometric_test.rb
@@ -167,16 +167,18 @@ class PostgresqlPointTest < ActiveRecord::PostgreSQLTestCase
 end
 
 class PostgresqlGeometricTest < ActiveRecord::PostgreSQLTestCase
+  include SchemaDumpingHelper
+
   class PostgresqlGeometric < ActiveRecord::Base; end
 
   setup do
     @connection = ActiveRecord::Base.connection
     @connection.create_table("postgresql_geometrics") do |t|
-      t.column :a_line_segment, :lseg
-      t.column :a_box, :box
-      t.column :a_path, :path
-      t.column :a_polygon, :polygon
-      t.column :a_circle, :circle
+      t.lseg    :a_line_segment
+      t.box     :a_box
+      t.path    :a_path
+      t.polygon :a_polygon
+      t.circle  :a_circle
     end
   end
 
@@ -232,6 +234,60 @@ class PostgresqlGeometricTest < ActiveRecord::PostgreSQLTestCase
 
     objs = PostgresqlGeometric.find_by_sql "SELECT isclosed(a_path) FROM postgresql_geometrics ORDER BY id ASC"
     assert_equal [false, true], objs.map(&:isclosed)
+  end
+
+  def test_schema_dumping
+    output = dump_table_schema("postgresql_geometrics")
+    assert_match %r{t\.lseg\s+"a_line_segment"$}, output
+    assert_match %r{t\.box\s+"a_box"$}, output
+    assert_match %r{t\.path\s+"a_path"$}, output
+    assert_match %r{t\.polygon\s+"a_polygon"$}, output
+    assert_match %r{t\.circle\s+"a_circle"$}, output
+  end
+end
+
+class PostgreSQLGeometricLineTest < ActiveRecord::PostgreSQLTestCase
+  include SchemaDumpingHelper
+
+  class PostgresqlLine < ActiveRecord::Base; end
+
+  setup do
+    unless ActiveRecord::Base.connection.send(:postgresql_version) >= 90400
+      skip("line type is not fully implemented")
+    end
+    @connection = ActiveRecord::Base.connection
+    @connection.create_table("postgresql_lines") do |t|
+      t.line :a_line
+    end
+  end
+
+  teardown do
+    @connection.drop_table 'postgresql_lines', if_exists: true
+  end
+
+  def test_geometric_line_type
+    g = PostgresqlLine.new(
+      a_line: '{2.0, 3, 5.5}'
+    )
+    g.save!
+
+    h = PostgresqlLine.find(g.id)
+    assert_equal '{2,3,5.5}', h.a_line
+  end
+
+  def test_alternative_format_line_type
+    g = PostgresqlLine.new(
+      a_line: '(2.0, 3), (4.0, 6.0)'
+    )
+    g.save!
+
+    h = PostgresqlLine.find(g.id)
+    assert_equal '{1.5,-1,0}', h.a_line
+  end
+
+  def test_schema_dumping_for_line_type
+    output = dump_table_schema("postgresql_lines")
+    assert_match %r{t\.line\s+"a_line"$}, output
   end
 end